requirements.txt
streamlit
pandas
rdkit
Pillow
openpyxl
